+++ src/pages/LetterWritePage/components/index.tsx
@@ -1,0 +1,19 @@
+import AgeSlider from './AgeSlider';
+import ConcernSelect from './ConcernSelect';
+import GenderSelect from './GenderSelect';
+import LetterPaper from './LetterPaper';
+import LetterReceiverSelect from './LetterReceiverSelect';
+import LetterReceiverContainer from './LetterReceiverContainer';
+import LetterWriteBottom from './LetterWriteBottom';
+import LetterWriteContent from './LetterWriteContent';
+
+export {
+  AgeSlider,
+  ConcernSelect,
+  GenderSelect,
+  LetterPaper,
+  LetterReceiverSelect,
+  LetterReceiverContainer,
+  LetterWriteBottom,
+  LetterWriteContent,
+};
